+++ src/spaceObjects/missiles/hvli.cpp
@@ -6,6 +6,7 @@ REGISTER_MULTIPLAYER_CLASS(HVLI, "HVLI");
 HVLI::HVLI()
 : MissileWeapon("HVLI", MissileWeaponData::getDataFor(MW_HVLI))
 {
+    setRadarSignatureInfo(0.1, 0.0, 0.0);
 }
 
 void HVLI::hitObject(P<SpaceObject> object)
@@ -20,4 +21,5 @@ void HVLI::hitObject(P<SpaceObject> object)
     e->setSize(category_modifier * 20);
     e->setPosition(getPosition());
     e->setOnRadar(true);
+    setRadarSignatureInfo(0.0, 0.0, 0.1);
 }