+++ src/spaceObjects/missiles/nuke.cpp
@@ -6,6 +6,7 @@ REGISTER_MULTIPLAYER_CLASS(Nuke, "Nuke");
 Nuke::Nuke()
 : MissileWeapon("Nuke", MissileWeaponData::getDataFor(MW_Nuke))
 {
+    setRadarSignatureInfo(0.0, 0.7, 0.1);
 }
 
 void Nuke::hitObject(P<SpaceObject> object)
@@ -18,4 +19,5 @@ void Nuke::hitObject(P<SpaceObject> object)
     e->setPosition(getPosition());
     e->setOnRadar(true);
     e->setExplosionSound("sfx/nuke_explosion.wav");
+    setRadarSignatureInfo(0.0, 0.7, 1.0);
 }