+++ src/spaceObjects/supplyDrop.h
@@ -18,8 +18,8 @@ public:
 
     virtual void collide(Collisionable* target, float force) override;
 
-    void setEnergy(float amount) { energy = amount; }
-    void setWeaponStorage(EMissileWeapons weapon, int amount) { if (weapon != MW_None) weapon_storage[weapon] = amount; }
+    void setEnergy(float amount); 
+    void setWeaponStorage(EMissileWeapons weapon, int amount);
 
     void onPickUp(ScriptSimpleCallback callback);